bugfix(react-dialog): fix scroll locking issues introduced by a regression
