+++ change/@fluentui-react-dialog-f8bedb1e-8ca8-4bd7-86ed-0c7623fd0368.json
@@ -1,0 +1,7 @@
+{
+  "type": "patch",
+  "comment": "bugfix: fix scroll locking issues introduced by a regression",
+  "packageName": "@fluentui/react-dialog",
+  "email": "bernardo.sunderhus@gmail.com",
+  "dependentChangeType": "patch"
+}
