+++ packages/react-components/react-dialog/src/components/DialogSurface/useDialogSurface.ts
@@ -5,11 +5,13 @@ import {
   isResolvedShorthand,
   slot,
   getIntrinsicElementProps,
+  useIsomorphicLayoutEffect,
 } from '@fluentui/react-utilities';
 import type { DialogSurfaceElement, DialogSurfaceProps, DialogSurfaceState } from './DialogSurface.types';
 import { useDialogContext_unstable } from '../../contexts';
 import { Escape } from '@fluentui/keyboard-keys';
 import { useDialogTransitionContext_unstable } from '../../contexts/dialogTransitionContext';
+import { useDisableBodyScroll } from '../../utils/useDisableBodyScroll';
 
 /**
  * Create the state required to render DialogSurface.
@@ -31,6 +33,7 @@ export const useDialogSurface_unstable = (
   const dialogRef = useDialogContext_unstable(ctx => ctx.dialogRef);
   const requestOpenChange = useDialogContext_unstable(ctx => ctx.requestOpenChange);
   const dialogTitleID = useDialogContext_unstable(ctx => ctx.dialogTitleId);
+  const open = useDialogContext_unstable(ctx => ctx.open);
 
   const handledBackdropClick = useEventCallback((event: React.MouseEvent<HTMLDivElement>) => {
     if (isResolvedShorthand(props.backdrop)) {
@@ -70,6 +73,20 @@ export const useDialogSurface_unstable = (
   if (backdrop) {
     backdrop.onClick = handledBackdropClick;
   }
+
+  const { disableBodyScroll, enableBodyScroll } = useDisableBodyScroll();
+  const isBodyScrollLocked = Boolean(open && modalType !== 'non-modal');
+  useIsomorphicLayoutEffect(() => {
+    if (isNestedDialog) {
+      return;
+    }
+    if (open && isBodyScrollLocked && transitionStatus === 'entering') {
+      disableBodyScroll();
+    } else if (transitionStatus === 'exited') {
+      enableBodyScroll();
+    }
+  }, [disableBodyScroll, enableBodyScroll, isBodyScrollLocked, isNestedDialog, open, transitionStatus]);
+
   return {
     components: { backdrop: 'div', root: 'div' },
     backdrop,
